requirements.txt
Django
djangorestframework
numpy
pandas
scipy
gunicorn
yfinance
